fix Travis
