+++ atf_msgs/CMakeLists.txt
@@ -16,10 +16,10 @@ add_message_files(
         Api.msg
         AtfResult.msg
         DataStamped.msg
+        Groundtruth.msg
         IO.msg
         Interface.msg
         InterfaceItem.msg
-        Groundtruth.msg
         KeyValue.msg
         MetricResult.msg
         Network.msg
