+++ atf_plotter/scripts/plot.py
@@ -13,6 +13,8 @@ from matplotlib.lines import Line2D
 
 import yaml
 
+import rosbag
+
 import pprint
 import re
 
@@ -30,175 +32,6 @@ import time
 
 MetricAggregate = namedtuple('MetricAggregate', 'metric_name testblock_name num min max mean median std data_points')
 
-
-class TimestampBoundData(object):
-    def __init__(self, data_dict):
-        self._data_dict = data_dict
-
-    @property
-    def timestamp(self):
-        stamp = self._data_dict['stamp']
-        secs = stamp['secs']
-        nsecs = stamp['nsecs']
-        return secs * nsecs * 1e-9
-
-    @property
-    def data(self):
-        return self._data_dict['data']
-
-
-class BaseYamlStructure(object):
-    def __init__(self, data_dict, parent=None):
-        self._data_dict = data_dict
-        self._sublevel_class = dict
-        self._sublevel_key = 'results'
-        self._sublevel_identifier = 'name'
-        self.parent = parent
-
-    @property
-    def sublevel_item_list(self):
-        return self._data_dict.get(self._sublevel_key, None)
-
-    @property
-    def sublevel_names(self):
-        return [sub_item.get(self._sublevel_identifier, None) for sub_item in self.sublevel_item_list]
-
-    @property
-    def name(self):
-        return self._data_dict.get('name', None)
-
-    @property
-    def groundtruth_result(self):
-        return self._data_dict.get('groundtruth_result', None)
-
-    @property
-    def groundtruth_error_message(self):
-        return self._data_dict.get('groundtruth_error_message', None)
-
-    def __contains__(self, item):
-        return item in self.sublevel_names
-
-    def __getitem__(self, key):
-        if key not in self:
-            raise KeyError
-
-        if self.sublevel_names.count(key) == 1:
-            ridx = self.sublevel_names.index(key)
-            sublevel_item = self._sublevel_class(self.sublevel_item_list[ridx], parent=self)
-            return sublevel_item
-        else:
-            sol = [sublevel_item for sublevel_item in self if sublevel_item.name == key]
-            return sol
-        #print 'count', self.sublevel_names.count(key)
-
-    def __iter__(self):
-       for sublevel_item in self.sublevel_item_list:
-          yield self._sublevel_class(sublevel_item, parent=self)
-
-    def __len__(self):
-        return len(self.sublevel_item_list)
-
-
-
-class Metric(BaseYamlStructure):
-    def __init__(self, data_dict, **kwargs):
-        BaseYamlStructure.__init__(self, data_dict=data_dict, **kwargs)
-        self._sublevel_key = 'details'
-        self._sublevel_identifier = 'key'
-
-    @property
-    def started(self):
-        return self._data_dict.get('started', None)
-
-    @property
-    def finished(self):
-        return self._data_dict.get('finished', None)
-
-    @property
-    def data(self):
-        return self._data_dict.get('data', None).get('data', None)
-
-    @property
-    def series(self):
-        raw = self._data_dict.get('series', None)
-        if raw is None or not len(raw):
-            return raw
-
-        return map(TimestampBoundData, raw)
-
-        #return [(tbd.timestamp, tbd.data) for tbd in map(TimestampBoundData, raw)]
-
-        #return self._data_dict.get('series', None)
-
-    @property
-    def series_tuple(self):
-        return [(tbd.timestamp, tbd.data) for tbd in self.series]
-
-    @property
-    def groundtruth(self):
-        return self._data_dict.get('groundtruth', None)
-
-    @property
-    def groundtruth_epsilon(self):
-        return self._data_dict.get('groundtruth_epsilon', None)
-
-    def get_detail(self, name):
-        return self[name]['value']
-
-    @property
-    def details(self):
-        return self._data_dict.get('details', None)
-
-
-
-class TestBlock(BaseYamlStructure):
-    def __init__(self, data_dict, **kwargs):
-        BaseYamlStructure.__init__(self, data_dict=data_dict, **kwargs)
-        self._sublevel_class = Metric
-
-
-
-class SingleTest(BaseYamlStructure):
-    def __init__(self, data_dict, **kwargs):
-        BaseYamlStructure.__init__(self, data_dict=data_dict, **kwargs)
-        self._sublevel_class = TestBlock
-
-    @property
-    def robot(self):
-        return self._data_dict.get('robot', None)
-
-    @property
-    def env(self):
-        return self._data_dict.get('env', None)
-
-    @property
-    def test_config(self):
-        return self._data_dict.get('test_config', None)
-
-    @property
-    def testblockset(self):
-        return self._data_dict.get('testblockset', None)
-
-    @property
-    def repetition_number(self):
-        num = int(self.name.rsplit('_', 1)[-1])
-        return num
-
-    @property
-    def test_case_ident(self):
-        num = self.name.rsplit('_', 1)[0]
-        return num
-
-
-
-class AtfResult(BaseYamlStructure):
-    def __init__(self, data_dict):
-        BaseYamlStructure.__init__(self, data_dict=data_dict)
-        self._sublevel_class = SingleTest
-
-
-
-
 class DemoPlotter(object):
 
     def __init__(self):
@@ -206,376 +39,193 @@ class DemoPlotter(object):
 
 
     def load_atf_result(self, filename):
+        
+        bag = rosbag.Bag(filename)
+        print "number of files in bag:", bag.get_message_count()
+        assert bag.get_message_count() == 1 # check if only one message is in the bag file
+            
+        for topic, msg, t in bag.read_messages():
+            #print type(msg), type(AtfResult()), isinstance(msg, AtfResult) # check if message is an AtfResult message
+            #assert isinstance(msg, AtfResult) # check if message is an AtfResult message
+            if topic == "atf_result":
+                self.atf_result = msg
+            else:
+                print "ERROR: invalid topic name in bag. ATF only expects topic 'atf_result'"
+            bag.close()
+            break
 
-        with open(filename, 'r') as f:
-            yaml_data = yaml.load(f, Loader=yaml.CLoader)
-            print "yaml_data=", yaml_data
+    def get_sorted_plot_dicts(self, atf_result, filter_tests, filter_testblocks, filter_metrics):
+        tbm = {}
+        tmb = {}
+        bmt = {}
+        mbt = {}
 
-        self.atf_result = AtfResult(yaml_data)
-
-    def _quicktest(self):
-        assert self.atf_result is not None
-
-        lfr = self.atf_result.sublevel_names[0]
-
-        cre = re.compile('^ts\d+_c\d+_r\d+_e\d+_s\d+_\d+$')
-        assert cre.match(lfr) is not None
-
-        assert lfr in self.atf_result
-
-
-    def get_flat_list_of_metrics(self):
-        combined_metrics = list()
-        for single_test in self.atf_result:
-            for test_block in single_test:
-                for metrics in test_block:
-                    combined_metrics.append(metrics)
-        return combined_metrics
-
-    def group_metric_list_by_repetition_number_for_testcase(self, test_case_ident):
-        grouped = dict()
-
-        combined_metrics = self.get_flat_list_of_metrics()
-        print "combined_metrics=", combined_metrics
-
-        for metric in combined_metrics:
-            if test_case_ident != metric.parent.parent.test_case_ident:
+        for test in atf_result.results:
+            #print test.name
+            if len(filter_tests) != 0 and test.name not in filter_tests:
                 continue
-
-            rn = metric.parent.parent.repetition_number
-            if rn not in grouped:
-                grouped[rn] = list()
-            grouped[rn].append(metric)
-
-        return grouped
-
-
-
-    def aggregate_grouped_metrics(self, grouped_metrics):
-        ilen = map(len, grouped_metrics.values())
-        assert areAllListValuesEqual(ilen)
-
-        unsuitable_metric_ident_detail_keys = [
-            'rot_first_euler',
-            'rot_last_euler',
-            'lin_first',
-            'lin_last',
-        ]
-
-        aggregated_metrics = list()
-        for aggregate_list in zip(*grouped_metrics.values()):
-
-            metric_ident = zip(*[(
-                m.name,
-                m.parent.name,
-                '/'.join(['%r' % d['value'] for d in m.details if d['key'] not in unsuitable_metric_ident_detail_keys])
-            ) for m in aggregate_list])
-
-            allEqual = areAllListValuesEqual(map(areAllListValuesEqual, metric_ident))
-            assert allEqual and True
-
-            metric_ident_reduced = [mi[0] for mi in metric_ident]
-
-            metric_values = zip(*[(
-                m.data,
-                m.groundtruth,
-                m.groundtruth_epsilon,
-                m.groundtruth_result
-            ) for m in aggregate_list])
-
-            metric_data = metric_values[0]
-
-            aggregate = MetricAggregate(
-                metric_name=metric_ident_reduced[0],
-                testblock_name=metric_ident_reduced[1],
-                num=len(metric_data),
-                min=min(metric_data),
-                max=max(metric_data),
-                mean=np.mean(metric_data),
-                median=np.median(metric_data),
-                std=np.std(metric_data),
-                data_points=metric_data
-            )
-
-            aggregated_metrics.append(aggregate)
-
-        return aggregated_metrics
-
-
-    def plot_data_and_groundtruth_for_given_metric_testblock_test(self, metric, testblock, test):
-        single_test = self.atf_result[test]
-        assert isinstance(single_test, SingleTest)
-
-        print 'single_test.repetition_number', single_test.repetition_number
-
-
-
-        test_block = single_test[testblock]
-        assert isinstance(test_block, TestBlock)
-
-        print "metric11111111111111=", metric
-
-        print "type(test_block)=", type(test_block)
-        print "test_block=", test_block
-        print "test_block[metric]=", test_block[metric]
-
-        metrics = test_block[metric]
-        print "type(metrics2323232323)=", type(metrics)
-        print "metrics23232323=", metrics
-        if type(metrics) != list: # make sure we handle a list (e.g. if only one tf_length_translation is configured, or if time metric is used)
-            metrics = [metrics]
-        #assert isinstance(metric, list)
-
-
-        fig, ax = plt.subplots()
-
-        width = 0.2
-
-
-
-
-        for idx, metric in enumerate(metrics):
-            #print metric.details
-            #label = metric['measured_frame']['value']
-            #label = '%s/%s/%s' % (single_test.name, test_block.name, metric.get_detail('measured_frame'))
-
-            print "metric2222=", metric
-            label = build_tmp_label(metric)
-            rects = ax.bar(idx - width / 2.0, metric.data, width, label=label)
-            autolabel_ax_bar(ax, rects)
-
-            print metric.data
-            print metric.groundtruth
-            print metric.groundtruth_epsilon
-            print
-
-        ax.legend()
-        plt.show()
-
-
-    def plot_all_metrics_for_given_testblock_test(self, testblock, test):
-        single_test = self.atf_result[test]
-        assert isinstance(single_test, SingleTest)
-
-        test_block = single_test[testblock]
-        assert isinstance(test_block, TestBlock)
-
-
-        fig, ax = plt.subplots()
-
-        width = 0.2
-
-        clut = cm._generate_cmap('jet', len(test_block))
-        for idx, metric in enumerate(test_block):
-            label = build_tmp_label(metric)
-            rects = ax.bar(idx - width / 2.0, metric.data, width, label=label, color=clut(idx))
-            autolabel_ax_bar(ax, rects)
-
-            print metric.data
-            print metric.groundtruth
-            print metric.groundtruth_epsilon
-            print
-
-        ax.legend()
-        plt.show()
-
-
-    def plot_all_metrics_testblocks_tests(self):
-        combined_metrics = self.get_flat_list_of_metrics()
-
-        fig, ax = plt.subplots()
-        width = 0.75
-
-        clut = cm._generate_cmap('jet', len(combined_metrics))
-
-        for idx, metric in enumerate(combined_metrics):
-            label = build_tmp_label(metric)
-            print label
-            print ' '*7, metric.details
-
-            rects = ax.bar(idx - width / 2.0, metric.data, width, label=label, color=clut(idx))
-            autolabel_ax_bar(ax, rects)
-
-        ax.legend()
-        plt.show()
-
-
-    def plot_aggregated_data_for_all_test_repetitions_for_given_test_ident(self, test_case_ident):
-        grm = self.group_metric_list_by_repetition_number_for_testcase(test_case_ident)
-        print "grm=", grm
-        amr = self.aggregate_grouped_metrics(grm)
-        print "amr=", amr
-
-        data = [am.data_points for am in amr]
-        labels = [am.metric_name for am in amr]
-        category = [am.testblock_name for am in amr]
-
-        fig, (ax1, ax2) = plt.subplots(nrows=1, ncols=2)
-
-        meanlineprops = dict(linestyle='--', color='purple')
-        hdls1 = ax1.boxplot(data, showmeans=True, meanline=True, meanprops=meanlineprops, labels=labels, patch_artist=True)
-
-        uniqe_category_list = list(set(category))
-        clut = cm._generate_cmap('Dark2', len(uniqe_category_list))
-        colors = [clut(uniqe_category_list.index(cat)) for cat in category]
-
-        #colors = [clut(i) for i in range(50)]
-
-        for idx, label in enumerate(ax1.get_xmajorticklabels()):
-            label.set_rotation(90)
-            label.set_color(colors[idx])
-            print idx, label#, colors[idx]
-            #label.set_horizontalalignment("right")
-
-        # fill with colors
-
-
-
-        #colors = ['blue', 'pink', 'lightblue', 'lightgreen']*5
-        for patch, color in zip(hdls1['boxes'], colors):
-            patch.set_facecolor(color)
-
-        custom_lines = [Line2D([0], [0], color=c, lw=4) for c in set(colors)]
-        ax1.legend(custom_lines, uniqe_category_list, ncol=2)
-
-
-        ax1.grid(True)
-
-
-        plt.tight_layout()
-
-        #plt.xticks(range(len(labels)+1), labels, rotation='vertical')
-        #plt.subplots_adjust(bottom=0.15)
-
-
-        #ax.boxplot(amr[0].data_points, showmeans=True, meanline=True)
-        #ax.violinplot(amr[0].data_points, [2], points=60, widths=0.7, showmeans=True, showextrema=True, showmedians=True, bw_method='silverman')
-
-        #for idx, ma in enumerate(amr):
-        #    try:
-        #        ax.violinplot(ma.data_points, [idx], points=60, widths=0.7, showmeans=False, showextrema=True, showmedians=True, bw_method=0.5)
-        #    except:
-        #        pass
-            #break
-
-
-        #ax.legend()
-        plt.show()
-
-    def plot_fmw(self):
-        print "self.atf_result=", self.atf_result.groundtruth_error_message
-        return
-
-        ###
-        test_case_ident = "ts0_c0_r0_e0_s0"
-        ###
-        grm = self.group_metric_list_by_repetition_number_for_testcase(test_case_ident)
-        print "grm=", grm
-        amr = self.aggregate_grouped_metrics(grm)
-        print "amr=", amr
-
-        data = [am.data_points for am in amr]
-        labels = [am.metric_name for am in amr]
-        category = [am.testblock_name for am in amr]
-
-        fig, (ax1, ax2, ax3, ax4) = plt.subplots(nrows=1, ncols=4)
-
-        meanlineprops = dict(linestyle='--', color='purple')
-        hdls1 = ax1.boxplot(data, showmeans=True, meanline=True, meanprops=meanlineprops, labels=labels, patch_artist=True)
-
-        uniqe_category_list = list(set(category))
-        clut = cm._generate_cmap('Dark2', len(uniqe_category_list))
-        colors = [clut(uniqe_category_list.index(cat)) for cat in category]
-
-        #colors = [clut(i) for i in range(50)]
-
-        for idx, label in enumerate(ax1.get_xmajorticklabels()):
-            label.set_rotation(90)
-            label.set_color(colors[idx])
-            print idx, label#, colors[idx]
-            #label.set_horizontalalignment("right")
-
-        # fill with colors
-
-
-
-        #colors = ['blue', 'pink', 'lightblue', 'lightgreen']*5
-        for patch, color in zip(hdls1['boxes'], colors):
-            patch.set_facecolor(color)
-
-        custom_lines = [Line2D([0], [0], color=c, lw=4) for c in set(colors)]
-        ax1.legend(custom_lines, uniqe_category_list, ncol=2)
-
-
-        ax1.grid(True)
-
-
-        plt.tight_layout()
-
-        #plt.xticks(range(len(labels)+1), labels, rotation='vertical')
-        #plt.subplots_adjust(bottom=0.15)
-
-
-        #ax.boxplot(amr[0].data_points, showmeans=True, meanline=True)
-        #ax.violinplot(amr[0].data_points, [2], points=60, widths=0.7, showmeans=True, showextrema=True, showmedians=True, bw_method='silverman')
-
-        #for idx, ma in enumerate(amr):
-        #    try:
-        #        ax.violinplot(ma.data_points, [idx], points=60, widths=0.7, showmeans=False, showextrema=True, showmedians=True, bw_method=0.5)
-        #    except:
-        #        pass
-            #break
-
-
-        #ax.legend()
-        plt.show()
-
-
-    def print_structure(self):
-        for test in self.atf_result:
-            print test.name
-            for testblock in test:
-                print ' '*3, testblock.name
-                old_name = None
-                metric_names = testblock.sublevel_names
-                series_len = 0
-                for metric in testblock:
-                    series_len += len(metric.series)
-                    if old_name == metric.name:
+            for testblock in test.results:
+                #print "  -", testblock.name
+                if len(filter_testblocks) != 0 and testblock.name not in filter_testblocks:
+                    continue
+
+                for metric in testblock.results:
+                    #print "    -", metric.name
+                    split_name = metric.name.split("::")
+                    if len(filter_metrics) != 0 and metric.name not in filter_metrics and split_name[0] not in filter_metrics:
                         continue
-                    old_name = metric.name
-                    print ' '*7, '%dx' % metric_names.count(metric.name), metric.name, '   (series available, length %d)' % series_len if series_len else ''
 
+                    # tbm
+                    if test.name                 not in tbm.keys():
+                        tbm[test.name] = {}
+                    if testblock.name            not in tbm[test.name].keys():
+                        tbm[test.name][testblock.name] = {}
+                    tbm[test.name][testblock.name][metric.name] = metric
 
+                    # tmb
+                    if test.name                 not in tmb.keys():
+                        tmb[test.name] = {}
+                    if metric.name               not in tmb[test.name].keys():
+                        tmb[test.name][metric.name] = {}
+                    tmb[test.name][metric.name][testblock.name] = metric
 
-def build_tmp_label(metric):
-    #label = '%s/%s/%s/%s' % (metric.parent.parent.name, metric.parent.name, metric.name, metric.get_detail('measured_frame'))
-    print "here-----------------------"
-    print "metric=", metric
-    label = '%s/%s/%s' % (metric.parent.parent.name, metric.parent.name, metric.name)
-    print "label=", label
-    try:
-        label += '/%s' % metric.get_detail('measured_frame')
-    except KeyError:
+                    # bmt
+                    if testblock.name            not in bmt.keys():
+                        bmt[testblock.name] = {}
+                    if metric.name               not in bmt[testblock.name].keys():
+                        bmt[testblock.name][metric.name] = {}
+                    bmt[testblock.name][metric.name][test.name] = metric
+
+                    # mbt
+                    if metric.name            not in mbt.keys():
+                        mbt[metric.name] = {}
+                    if testblock.name         not in mbt[metric.name].keys():
+                        mbt[metric.name][testblock.name] = {}
+                    mbt[metric.name][testblock.name][test.name] = metric
+
+        ret = {}
+        ret['tbm'] = tbm
+        ret['tmb'] = tmb
+        ret['bmt'] = bmt
+        ret['mbt'] = mbt
+        return ret
+
+    def merge(self):
         pass
 
-    return label
+    def print_structure(self):
+        for test in self.atf_result.results:
+            print test.name
+            for testblock in test.results:
+                print "  -", testblock.name
+                for metric in testblock.results:
+                    print "    -", metric.name
 
+    def plot_fmw(self, style, filter_tests, filter_testblocks, filter_metrics):
 
-def areAllListValuesEqual(lst):
-    return not lst or lst.count(lst[0]) == len(lst)
+        sorted_atf_results = self.get_sorted_plot_dicts(self.atf_result, filter_tests, filter_testblocks, filter_metrics)
 
+        if style not in sorted_atf_results.keys():
+            print "ERROR: style '%s' not implemented"%style
+            return
+        plot_dict = sorted_atf_results[style]
 
-def autolabel_ax_bar(ax, rects):
-    """Attach a text label above each bar in *rects*, displaying its height."""
-    for rect in rects:
-        height = rect.get_height()
-        ax.annotate('{}'.format(height),
-                    xy=(rect.get_x() + rect.get_width() / 2, height),
-                    xytext=(0, 3),  # 3 points vertical offset
-                    textcoords="offset points",
-                    ha='center', va='bottom')
+        rows = plot_dict.keys()
+        cols = []
+        plots = []
+        nr_unique_plots = 0
+        for row in rows:
+            cols_tmp = plot_dict[row].keys()
+            for col in cols_tmp:
+                if col not in cols:
+                    cols.append(col)
+                    for plot in plot_dict[row][col].keys():
+                        if plot not in plots:
+                            plots.append(plot)
 
+        # sort alphabetically
+        rows.sort()
+        cols.sort()
+        plots.sort()
+
+        print "\nplotting in style '%s' (rows: %d, cols: %d, plots: %d)"%(style, len(rows), len(cols), len(plots))
+        meanlineprops = dict(linestyle='--', color='purple')
+        fig, axs = plt.subplots(len(rows), len(cols), sharex=True, figsize=(20, 15)) # FIXME calculate width with nr_testblocks
+
+        # always make this a numpy 2D matrix to access rows and cols correclty if len(rows)=1 or len(cols)=1
+        axs = np.atleast_2d(axs) 
+        axs = axs.reshape(len(rows), len(cols)) 
+
+        for row in rows:
+            #print "\nrow=", row
+            
+            
+            for col in cols:
+                #print "  col=", col
+
+                x = np.arange(len(plots))
+                ax = axs[rows.index(row)][cols.index(col)]
+
+                # format x axis
+                ax.set_xticks(x)
+                ax.set_xticklabels(plots)
+                ax.set_xlim(-1, len(plots))
+
+                y_min = 0
+                y_max = 0
+
+                # only set title for upper row and ylabel for left col
+                if rows.index(row) == 0:
+                    ax.set_title(col)
+                if cols.index(col) == 0:
+                    ax.set_ylabel(row, rotation=90)
+
+                for plot in plots:
+                    #print "    plot=", plot
+                    try:
+                        metric_result = plot_dict[row][col][plot]
+                    except KeyError:
+                        #print "skip", row, col, plot
+                        continue
+
+                    ax.grid(True)
+                    nr_unique_plots += 1
+
+                    data = metric_result.data.data
+                    lower = metric_result.groundtruth.data - metric_result.groundtruth.epsilon
+                    upper = metric_result.groundtruth.data + metric_result.groundtruth.epsilon
+                    y_min = min(-0.1, data, lower, upper)
+                    y_max = max(data, lower, upper)
+
+                    yerr = [[0], [0]]
+                    if metric_result.groundtruth.available:
+                        yerr = [[data - lower], [upper - data]]
+                    ax.errorbar(plots.index(plot), data, yerr=yerr, fmt='D', markersize=12)
+
+                    # plot min and max
+                    ax.plot(plots.index(plot), metric_result.min.data, '^', markersize=8)
+                    ax.plot(plots.index(plot), metric_result.max.data, 'v', markersize=8)
+
+                # format y axis
+                (y_min_auto, y_max_auto) = ax.get_ylim()
+                y_min = min (y_min_auto, y_min)
+                y_max = max (y_max_auto, y_max)
+                ax.set_ylim(y_min - 0.2*abs(y_min), y_max + 0.2*abs(y_max)) # make it a little bigger than the min/max values
+
+        fig.autofmt_xdate(rotation=45)
+        plt.tight_layout()
+
+        title = "ATF Result for %s\ntotal # of tests: %d\ntotal # of plots: %d"%("PACKAGE", len(self.atf_result.results), nr_unique_plots)   # replace PACKAGE with self.atf_result.package (needs to be added to message first)
+        st = fig.suptitle(title, fontsize="x-large")
+        # shift subplots down:
+        st.set_y(0.95)
+        fig.subplots_adjust(top=0.85)
+
+        fig.savefig("/tmp/test.png")
+        plt.show()
+
+        print "DONE plot_fmw"
+        return
 
 if __name__ == '__main__':
     # example call could be:
@@ -583,10 +233,11 @@ if __name__ == '__main__':
     # to get info about the file, this could be helpful:
     #   rosrun atf_plotter plot.py info-structure ~/atf_result.txt
 
-    add_metric = lambda sp: sp.add_argument('--metric', '-m', type=str, dest='metric', required=True, help='TBD')
-    add_testblock = lambda sp: sp.add_argument('--testblock', '-tb', type=str, dest='testblock', required=True, help='TBD')
-    add_test = lambda sp: sp.add_argument('--test', '-t', type=str, dest='test', required=True, help='TBD')
+    add_test = lambda sp: sp.add_argument('--test', '-t', type=str, dest='test', default="", help='TBD')
     add_test_case_ident = lambda sp: sp.add_argument('--testident', '-ti', type=str, dest='test_case_ident', required=True, help='like test name without repetition, e.g. ts0_c0_r0_e0_s0')
+    add_testblock = lambda sp: sp.add_argument('--testblock', '-tb', type=str, dest='testblock', default="", help='TBD')
+    add_metric = lambda sp: sp.add_argument('--metric', '-m', type=str, dest='metric', default="", help='TBD')
+    add_style = lambda sp: sp.add_argument('--style', '-s', type=str, dest='style', default='bmt', help='style, e.g. tbm (default) tmb, bmt, ...')
 
 
     parser = argparse.ArgumentParser(
@@ -623,6 +274,10 @@ if __name__ == '__main__':
     add_test_case_ident(sub_parser)
 
     sub_parser = subparsers.add_parser('plot-fmw', help='fmw test plot')
+    add_style(sub_parser)
+    add_test(sub_parser)
+    add_testblock(sub_parser)
+    add_metric(sub_parser)
 
     sub_parser = subparsers.add_parser('compare-a', help='visualize comparison for a given metric in various testblocks of a given test, e.g. path_length in testblock testblock_small and testblock testblock_large from atf_test/ts0_c0_r0_e0_s0_0')
 
@@ -704,7 +359,7 @@ if __name__ == '__main__':
     sys.stdout.flush()
     #dp._quicktest()
 
-    #dp.print_structure()
+    dp.print_structure()
 
 
     if argparse_result.command == 'plot-metric':
@@ -723,7 +378,7 @@ if __name__ == '__main__':
     elif argparse_result.command == 'plot-d':
         dp.plot_aggregated_data_for_all_test_repetitions_for_given_test_ident(argparse_result.test_case_ident)
     elif argparse_result.command == 'plot-fmw':
-        dp.plot_fmw()
+        dp.plot_fmw(argparse_result.style, argparse_result.test, argparse_result.testblock, argparse_result.metric)
     elif argparse_result.command == 'info-structure':
         dp.print_structure()
     else:
